# contracts/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Self-hosted SRE/observability aggregator core API.",
    "title": "sre-kit API",
    "contact": {},
    "version": "1.0"
  },
  "paths": {
...
    "/api/checks": {
      "get": {
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "description": "Query current check statuses",
        "tags": [
          "telemetry"
        ],
        "summary": "List checks",
        "parameters": [
          {
            "description": "source ID filter",
            "name": "source",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/internal_telemetry_interfaces_http.checkResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/events": {
      "get": {
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "description": "Query the event feed",
        "tags": [
          "telemetry"
        ],
        "summary": "List events",
        "parameters": [
          {
            "description": "source ID filter",
            "name": "source",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "max events returned",
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/internal_telemetry_interfaces_http.eventResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sources/{id}/ingest-token": {
      "post": {
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "telemetry"
        ],
        "summary": "Rotate a Source push token",
        "parameters": [
          {
            "description": "source ID",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sources/{id}/records": {
      "post": {
        "tags": [
          "telemetry"
        ],
        "summary": "Push an idempotent telemetry batch",
        "parameters": [
          {
            "description": "source ID",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "unique batch key",
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/internal_telemetry_interfaces_http.pushBatch"
              }
            }
          },
          "description": "versioned Metric/Check/Event batch",
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/stream": {
      "get": {
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "description": "WebSocket: subscribe/unsubscribe by source_id, receive live Metric/Check/Event frames",
        "tags": [
          "telemetry"
        ],
        "summary": "Live stream",
        "responses": {}
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "name": "session",
        "in": "cookie"
      }
    },
    "schemas": {
...
      "internal_telemetry_interfaces_http.checkResponse": {
        "type": "object",
        "properties": {
          "meta": {
            "type": "object",
            "additionalProperties": {}
          },
          "name": {
            "type": "string"
          },
          "source_id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "ts": {
            "type": "string"
          }
        }
      },
      "internal_telemetry_interfaces_http.eventResponse": {
        "type": "object",
        "properties": {
          "labels": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "level": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "source_id": {
            "type": "string"
          },
          "ts": {
            "type": "string"
          }
        }
      },
...
      "internal_telemetry_interfaces_http.pushBatch": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/internal_telemetry_interfaces_http.pushRecord"
            }
          },
          "schema_version": {
            "type": "string"
          }
        }
      },
      "internal_telemetry_interfaces_http.pushRecord": {
        "type": "object",
        "properties": {
          "labels": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "level": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "meta": {
            "type": "object",
            "additionalProperties": {}
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "value": {
            "type": "number"
          }
        }
      }
    }
  }
}